// src/app/models/v_MovementsDetail.model.ts
export class ViewMovementsDetailsModel{
    idMovementDetail: number;
    idMovement:number;
    dateMov: Date;
    taxAdd: number;
    discount: number;
    multiplier: number;
    codeProduct: string;
    descriptionProduct: string;
    descriptionWarehouse: string;
    quantity: number;
    value: number;
    valorBruto: number;
    valorImpuesto:number;
    valorDescuento: number;
    total: number;
    idUEN: number;
    uEN: string;
    tMovement: string;
}